v2.1 - August 18th, 2015
v2.1 - August 18th, 2015
o Added support for responsive Facebook videos to the mod.
o Added option to control the default width of the post when it is not specified.
o Added option to control the default width of the video when it is not specified.
o Updated Greek UTF8 language modifications, due to several new strings to support new options.
o Moved text for mod options from "index" to "ManageSettings" language file.